Added Statistic Calculator
