+++ calculators/src/App.js
@@ -10,7 +10,6 @@ import MainOctToBinAndBinToOct from './computer_science/OctToBinAndBinToOct/Main
 import MainBinaryAdder from './computer_science/BinaryAdder/MainBinaryAdder';
 import MainOctalAdder from './computer_science/OctalAdder/MainOctalAdder';
 import MainHexadecimalAdder from './computer_science/HexadecimalAdder/MainHexadecimalAdder';
-import StatisticsCalculator from './mathematical/Statistics/StatisticsCalculator';
 import MainStatistics from './mathematical/statistics/MainStatistics';
 import MainBMI from './health/BMI/MainBMI';
 import MainBMR from './health/BMR/MainBMR';
@@ -33,7 +32,6 @@ function App() {
             {/* Mathematical */}
             <Route path="simpleinterest" element={<MainSI />} />
             <Route path="compoundinterest" element={<MainCI />} />
-            <Route path="statistics" element={<StatisticsCalculator />} />
             <Route path="statistics" element={<MainStatistics />} />
             {/* Health */}
             <Route path="bmi" element={<MainBMI />} />
